\docType{package}
\name{bfa-package}
\alias{bfa-package}
\title{Bayesian Factor Analysis}
\description{
  Bayesian Factor Analysis
}
\details{
  \tabular{ll}{ Package: \tab bfa\cr Type: \tab Package\cr
  Version: \tab 1.0\cr Date: \tab 2011-06-03\cr License:
  \tab GPL-3\cr LazyLoad: \tab yes\cr }

  This package provides model fitting for several Bayesian
  factor models including Gaussian, ordinal probit, mixed
  and semiparametric copula factor models. Normal,
  continuous shrinkage and point-mass sparsity priors for
  factor loadings are available. Currently only exploratory
  factor models are implemented.
}
\author{
  Jared Murray \email{jared.murray@stat.duke.edu}
}
\keyword{package}

