\name{biplot.bfa}
\alias{biplot.bfa}
\title{Display a biplot...}
\usage{\method{biplot}{bfa}(x, factors=c(1, 2), ...)
}
\description{Display a biplot}
\value{Shows a biplot}
\arguments{\item{x}{A bfa object}
\item{factors}{Numeric/integer vector giving indices of the factors to plot}
\item{...}{Additional arguments to biplot; see \code{?biplot}}
}

