\name{corr.samp}
\alias{corr.samp}
\title{Compute samples of the correlation matrix}
\usage{
  corr.samp(model)
}
\arguments{
  \item{model}{\code{bfa} model object}
}
\value{
  A p x p x (no. of mcmc samples) array containing samples
  of the correlation matrix
}
\description{
  Compute samples of the correlation matrix
}

