\name{mean.bfa}
\alias{mean.bfa}
\title{Extract posterior means from bfa object...}
\usage{\method{mean}{bfa}(x, ...)
}
\description{Extract posterior means from bfa object}
\value{A list with elements loadings and scores containing MCMC means}
\arguments{\item{x}{A bfa object}
\item{...}{Ignored}
}

