\name{ratings08}
\alias{ratings08}
\alias{ratings08}
\title{Special Interest Group Ratings (2008)...}
\usage{ratings08}
\description{Special Interest Group Ratings (2008)}
\details{This dataset consists of SIG ratings of members of the US House and Senate.
It contains two objects: \code{ratings}, a data frame with the ratings
information, and \code{groups} which contains information about the
special interest groups}
\docType{data}
\author{Jared Murray \email{jared.murray@stat.duke.edu}}

