\name{reg.samp}
\alias{reg.samp}
\title{Compute regression coefficients for variables in (index) against the rest}
\usage{
  reg.samp(model, index = c(1))
}
\arguments{
  \item{model}{\code{bfa} model object}

  \item{index}{index(es) of the response}
}
\description{
  Compute regression coefficients for variables in (index)
  against the rest
}

