\name{utri_restrict}
\alias{utri_restrict}
\title{Upper triangular loadings restriction...}
\usage{utri_restrict(p, k)
}
\description{Upper triangular loadings restriction}
\details{Returns the restriction matrix corresponding to Geweke & Zhou (1996)
upper-triangular positive diagonal condition
'}
\value{p by k matrix}
\arguments{\item{p}{number of rows (variables)}
\item{k}{number of columns (factors)}
}

