\name{var.bfa}
\alias{var.bfa}
\title{Extract posterior variances from sbfac object...}
\usage{\method{var}{bfa}(x, ...)
}
\description{Extract posterior variances from sbfac object}
\value{A list with elements loadings and scores containing MCMC variances}
\arguments{\item{x}{A bfa object}
\item{...}{Ignored}
}

