\docType{package}
\name{bfa-package}
\alias{bfa-package}
\title{Bayesian Factor Analysis}
\description{
  Bayesian Factor Analysis
}
\details{
  \tabular{ll}{ Package: \tab bfa\cr Type: \tab Package\cr
  Version: \tab 0.2\cr Date: \tab 2011-10-26\cr License:
  \tab GPL-3\cr LazyLoad: \tab yes\cr }

  This package provides model fitting for several Bayesian
  factor models including Gaussian, ordinal probit, mixed
  and semiparametric Gaussian copula factor models. Normal,
  generalized double Pareto and point-mass sparsity priors
  for factor loadings are available.
}
\author{
  Jared Murray \email{jared.murray@stat.duke.edu}
}
\keyword{package}

