\docType{data}
\name{ratings08}
\alias{ratings08}
\title{Special Interest Group Ratings (2008)}
\format{List of two data frames: \code{ratings} and \code{groups} (see above)}
\source{
  Scraped from Project VoteSmart:
  \url{http://www.projectvotesmart.org}
}
\usage{
  ratings08
}
\description{
  This dataset consists of SIG ratings of members of the US
  House and Senate. It contains two objects:
  \code{ratings}, a data frame with the ratings
  information, and \code{groups} which contains information
  about the special interest groups
}
\author{
  Jared Murray \email{jared.murray@stat.duke.edu}
}

